+++ src/plugin/extra/common.h
@@ -1,7 +1,7 @@
 #ifndef GCOMMON
 #define GCOMMON
 
-int sf_parse_info_lean (struct StataInfo *st_info);
+int sf_parse_info_lean (struct StataInfo *st_info, int level);
 
 void sf_free_lean (struct StataInfo *st_info);