+++ src/plugin/extra/hashsort.h
@@ -1,0 +1,6 @@
+#ifndef GHASHSORT
+#define GHASHSORT
+
+int sf_hashsort (struct StataInfo *st_info);
+
+#endif